feat: Addressed review comments
