+++ Makefile
@@ -3,14 +3,10 @@
 
 test:
 	poetry run python -m unittest discover -s supra_sdk/ -p '*.py' -t ..
-	poetry run python -m unittest test.t_endpoints
-	poetry run python -m unittest test.t_automation_txn
 	poetry run behave
 
 test-coverage:
 	poetry run python -m coverage run -m unittest discover -s supra_sdk/ -p '*.py' -t ..
-	poetry run python -m coverage run -m unittest test.t_endpoints
-	poetry run python -m coverage run -m unittest test.t_automation_txn
 	poetry run python -m coverage report
 
 test-spec:
@@ -29,7 +25,7 @@ examples:
 	poetry run python -m examples.supra_token
 	poetry run python -m examples.fee_payer_transfer_coin
 	poetry run python -m examples.rotate_key
-	# poetry run python -m examples.read_aggregator
+	poetry run python -m examples.read_aggregator
 	# poetry run python -m examples.secp256k1_ecdsa_transfer_coin
 	poetry run python -m examples.simple_supra_token
 	# poetry run python -m examples.simple_nft
@@ -38,22 +34,7 @@ examples:
 	poetry run python -m examples.transfer_two_by_two
 	poetry run python -m examples.multikey
 
-examples_cli:
-	@echo "These examples still need some work"
-# poetry run python -m examples.hello_blockchain
-# poetry run python -m examples.large_package_publisher CURRENTLY BROKEN -- OUT OF GAS
-# poetry run python -m examples.multisig
-# poetry run python -m examples.your_coin
-
 integration_test:
-	@echo "Disabled for now"
-# poetry run python -m unittest -b examples.integration_test
-
-#########
-# SUPRA #
-#########
-test-endpoints:
-	poetry run python -m unittest test.t_endpoints
-	poetry run python -m unittest test.t_automation_txn
+	poetry run python -m unittest -b examples.integration_test
 
 .PHONY: examples fmt lint test
