+++ examples/transaction_batching.py
@@ -15,14 +15,11 @@ from supra_sdk.account_address import AccountAddress
 from supra_sdk.account_sequence_number import AccountSequenceNumber
 from supra_sdk.async_client import ClientConfig, FaucetClient, RestClient
 from supra_sdk.bcs import Serializer
-from supra_sdk.supra_token_client import Property, PropertyMap, SupraTokenClient
+from supra_sdk.supra_token_client import (Property, PropertyMap,
+                                          SupraTokenClient)
 from supra_sdk.transaction_worker import TransactionWorker
-from supra_sdk.transactions import (
-    EntryFunction,
-    SignedTransaction,
-    TransactionArgument,
-    TransactionPayload,
-)
+from supra_sdk.transactions import (EntryFunction, SignedTransaction,
+                                    TransactionArgument, TransactionPayload)
 from supra_sdk.type_tag import StructTag, TypeTag
 
 from .common import FAUCET_URL, NODE_URL
