+++ examples/your_coin.py
@@ -151,8 +151,8 @@ async def main(moon_coin_path: str):
 
 
 if __name__ == "__main__":
-    assert len(sys.argv) == 2, (
-        "Expecting an argument that points to the moon_coin directory."
-    )
+    assert (
+        len(sys.argv) == 2
+    ), "Expecting an argument that points to the moon_coin directory."
 
     asyncio.run(main(sys.argv[1]))
